Modularizing duplicated code.
