+++ test/test_utils.py
@@ -1,0 +1,9 @@
+"""Tests for AugmentedNet.utils."""
+
+import unittest
+
+import AugmentedNet.utils
+
+
+class TestUtils(unittest.TestCase):
+    pass
